fix(harness): rev
