+++ src/extensions/ferment/ask-user.ts
@@ -29,7 +29,7 @@ import type { ExtensionAPI } from "@earendil-works/pi-coding-agent"
 import { renderLabeledSuccessCriteria } from "../../ferment/success-criteria.js"
 import type { Ferment, ScopingQuestionType } from "../../ferment/types.js"
 import { YES_NO_OPTIONS } from "../questionnaire-reducer.js"
-import { normalizeQuestionType } from "../questionnaire-shared.js"
+import { normalizeQuestionType } from "../questionnaire.js"
 import { type JudgeApiResult, judgeApiCall } from "./judge.js"
 import { promptForm } from "./prompt-ui.js"
 import type { FermentRuntime } from "./runtime.js"
