+++ src/extensions/ferment/nudge.test.ts
@@ -359,10 +359,6 @@ describe("scoping progress nudge", () => {
 			}),
 			{ triggerTurn: true },
 		)
-		const sentMessage = vi.mocked(pi.sendMessage).mock.calls[0][0] as { content: Array<{ text?: string }> }
-		const nudgeText = sentMessage.content[0]?.text ?? ""
-		expect(nudgeText).toContain("call ask_user")
-		expect(nudgeText).not.toContain("call questionnaire")
 	})
 
 	it("resets the counter when a scoping-progress tool is seen", () => {
