+++ src/extensions/questionnaire.ts
@@ -17,21 +17,13 @@
 
 import type { ExtensionAPI, ExtensionContext } from "@earendil-works/pi-coding-agent"
 import { Text, truncateToWidth } from "@earendil-works/pi-tui"
-import { Type } from "typebox"
+import { type Static, Type } from "typebox"
 
 import { createToolVisibility } from "./prompt-construction/tool-visibility.js"
 import { createQuestionForm } from "./questionnaire-form.js"
-import type { Answer, Question } from "./questionnaire-reducer.js"
-import {
-	formatAnswerText,
-	normalizeQuestion,
-	normalizeQuestionType,
-	validateQuestions,
-	validateRawQuestions,
-} from "./questionnaire-shared.js"
+import { type Answer, type Question, type QuestionType, YES_NO_OPTIONS } from "./questionnaire-reducer.js"
 
 export type { Answer, Question }
-export { formatAnswerText, normalizeQuestionType }
 
 interface QuestionnaireResult {
 	questions: Question[]
@@ -87,6 +79,25 @@ const QuestionnaireParams = Type.Object({
 
 export type QuestionnaireQuestionTypeInput = string
 
+/** Normalize an agent-supplied question type to the canonical vocabulary.
+ *  Only an omitted (undefined) type defaults to "single"; any other string must
+ *  be one of single|multi|text|confirm (case-insensitive). Unknown strings throw
+ *  rather than silently becoming "single" — there is one vocabulary, no aliases.
+ *  Normal tool calls never reach the throw because the TypeBox pattern rejects
+ *  bad types first; it guards direct/internal callers that bypass the schema. */
+export function normalizeQuestionType(type: string | undefined): QuestionType {
+	if (type === undefined) return "single"
+	const canonical: Record<string, QuestionType> = {
+		single: "single",
+		multi: "multi",
+		text: "text",
+		confirm: "confirm",
+	}
+	const mapped = canonical[type.toLowerCase()]
+	if (!mapped) throw new Error(`Unknown question type: "${type}". Expected single, multi, text, or confirm.`)
+	return mapped
+}
+
 function errorResult(
 	message: string,
 	questions: Question[] = [],
@@ -95,6 +106,71 @@ function errorResult(
 		content: [{ type: "text", text: message }],
 		details: { questions, answers: [], cancelled: true },
 	}
+}
+
+function normalizeQuestion(q: Static<typeof QuestionSchema>, index: number): Question {
+	const type = normalizeQuestionType(q.type)
+	const rawOptions = q.options ?? []
+	const normalizedOptions = rawOptions.map((opt) => ({
+		id: opt.id,
+		label: opt.label,
+		description: opt.description,
+	}))
+	return {
+		id: q.id,
+		label: q.label || `Q${index + 1}`,
+		prompt: q.prompt,
+		type,
+		options: type === "confirm" ? [...YES_NO_OPTIONS] : normalizedOptions,
+		allowOther: q.allowOther ?? type === "single",
+		required: q.required !== false,
+	}
+}
+
+function validateRawQuestions(questions: Static<typeof QuestionSchema>[]): string | undefined {
+	for (const q of questions) {
+		const type = normalizeQuestionType(q.type)
+		if (type !== "confirm") continue
+		if ((q.options?.length ?? 0) > 0) {
+			return `Question "${q.id}" is type "confirm" and must not have options — confirm is always Yes/No.`
+		}
+		if (q.allowOther) {
+			return `Question "${q.id}" is type "confirm" and must not set allowOther — confirm is always Yes/No.`
+		}
+	}
+	return undefined
+}
+
+function validateQuestions(questions: Question[]): string | undefined {
+	for (const q of questions) {
+		if ((q.type === "single" || q.type === "multi") && q.options.length === 0 && !q.allowOther) {
+			return `Question "${q.id}" is type "${q.type}" but has no options and allowOther is false.`
+		}
+	}
+	return undefined
+}
+
+/** Format answers as human-readable text for the LLM. */
+export function formatAnswerText(questions: Question[], answers: Answer[]): string {
+	return answers
+		.map((a) => {
+			const qLabel = questions.find((q) => q.id === a.id)?.label || a.id
+			if (a.values && a.labels) {
+				const items = a.labels
+					.map((l, i) => {
+						const idx = a.indices?.[i]
+						return idx ? `${idx}. ${l}` : l
+					})
+					.join(", ")
+				return `${qLabel}: user selected: ${items}`
+			}
+			if (a.wasCustom) {
+				return `${qLabel}: user wrote: ${a.label}`
+			}
+			const display = a.index ? `${a.index}. ${a.label}` : a.label
+			return `${qLabel}: user selected: ${display}`
+		})
+		.join("\n")
 }
 
 // ─── Extension ────────────────────────────────────────────────────────────────
